+++ CHANGELOG.md
@@ -10,9 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Show tiles cache size in region management UI
-- Online Nominatim geocoding
+- Online Photon geocoding
 - Zoom-in using double-tap
 - Map marker for search results
+- Map long-press touch action for searching entities at a specific location
 
 ### Changed